Sharing internal class of v2.00.0000 of a tool from 2008.
This is the flagship class object responsible for driving the CSS validator on my site. I have posted this here ahead of the entire CSS email validation project and plan to make a final update upon completion. I have also put to gitHub...version 1.0.0 of the CSS validator project from 2008; this is for historical reasons.

This is only one class out of many that comprise what will become v2.00.0000 of CRNRSTN ::.

CRNRSTN :: An Open Source PHP Class Library providing a robust services interface layer to both facilitate, augment, and enhance the operations of code base for an application across multiple hosting environments.